Alpha Functionality
We implemented the Spotify oAuth, so that we are able to access the user's information and it properly redirects us to the homepage. Additionally, we selectively display information depending on the status of the login. In the backend, we also have functionality configured to fetch the playlist data, and we are working on adding functionality for displaying the information that we have already fetched.